third_party/WebKit/LayoutTests/http/tests/storage/async_cookies/document_getAll_multiple.html

<!doctype html>
<meta charset="utf-8">
<title>Async Cookes: cookieStore.getAll() sees cookieStore.set() cookie</title>
<link rel="author" href="pwnall@chromium.org" title="Victor Costan">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');
  await cookieStore.set('cookie-name-2', 'cookie-value-2');
  await cookieStore.set('cookie-name-3', 'cookie-value-3');

  const cookies = await cookieStore.getAll();
  cookies.sort((a, b) => a.name.localeCompare(b.name));
  assert_equals(cookies.length, 3);
  assert_equals(cookies[0].name, 'cookie-name');
  assert_equals(cookies[0].value, 'cookie-value');
  assert_equals(cookies[1].name, 'cookie-name-2');
  assert_equals(cookies[1].value, 'cookie-value-2');
  assert_equals(cookies[2].name, 'cookie-name-3');
  assert_equals(cookies[2].value, 'cookie-value-3');

  try { await cookieStore.delete('cookie-name'); } catch (e) {}
  try { await cookieStore.delete('cookie-name-2'); } catch (e) {}
  try { await cookieStore.delete('cookie-name-3'); } catch (e) {}
}, 'cookieStore.getAll returns multiple cookies written by cookieStore.set');

</script>
